strands-agents/enterprise-ai/app-db/requirements.txt
sqlalchemy==2.0.23
psycopg2-binary>=2.9.0
python-dotenv==1.0.0
alembic==1.13.1
